Renames: remove "multimodal" from Gemini Live types
