+++ src/pipecat/services/gemini_live/__init__.py
@@ -1,0 +1,3 @@
+from .file_api import GeminiFileAPI
+from .gemini import GeminiLiveLLMService
+from .vertex import GeminiLiveVertexLLMService
